requirements.txt
# Core
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0

# Parsing
pandas>=2.0
openpyxl>=3.1
PyPDF2>=3.0
pdfplumber>=0.10

# Invoice verification
requests>=2.31
ddddocr>=1.5.0
onnxruntime>=1.16.0
opencv-python-headless>=4.8
Pillow>=10.0

# Browser automation (screenshot capture)
playwright>=1.40
